+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+### Changed
+
+ - Rename `S.ty` to `S.int_ty` (in accordance with `generic_int_ty`).
+
 ## 0.1.2 - 2024-05-15
 
 ### Added